requirement.txt
django-html5
django-allauth
pyyaml
tweetpony
